+++ packages/moonstone/VirtualList/GridListImageItem.js
@@ -112,8 +112,8 @@ const GridListImageItemBase = kind({
 						</div>
 					) : null
 				}
-				{caption ? (<MarqueeText marqueeOn="hover" className={css.caption}>{caption}</MarqueeText>) : null}
-				{subCaption ? (<MarqueeText marqueeOn="hover" className={css.subCaption}>{subCaption}</MarqueeText>) : null}
+				{caption ? (<MarqueeText marqueeOn="focus" className={css.caption}>{caption}</MarqueeText>) : null}
+				{subCaption ? (<MarqueeText marqueeOn="focus" className={css.subCaption}>{subCaption}</MarqueeText>) : null}
 			</Controller>
 		);
 	}